# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fingarden API Docs",
    "version": "1.0.0",
    "title": "Fingarden"
  },
  "basePath": "/",
  "paths": {
...
    "/tags": {
      "get": {
        "tags": ["tag"],
        "summary": "Get array of tags",
        "description": "",
        "operationId": "getTags",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "in": "query",
            "name": "name",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "url",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "parent",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "children",
            "required": false
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "in": "query",
            "name": "products",
            "required": false
          },
          {
            "type": "string",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "default": "name"
          },
          {
            "type": "string",
            "in": "query",
            "name": "orderBy",
            "required": false,
            "default": "DESC",
            "description": "can be ASC or DESC"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "type": "string",
            "in": "query",
            "name": "limit",
            "required": false,
            "default": 10,
            "description": "qty rows to take for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TagResponse"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tag"],
        "summary": "Create a tag",
        "description": "",
        "operationId": "createTag",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Tag request"
          },
          "403": {
            "description": "Forbidden"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/tags/{tagId}": {
      "get": {
        "tags": ["tag"],
        "summary": "Find tag by id",
        "operationId": "findTagById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "description": "ID of tag to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Tag not found"
          }
        }
      },
      "put": {
        "tags": ["tag"],
        "summary": "Update the tag",
        "description": "",
        "operationId": "updateTag",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "ID of tag to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "$ref": "#/definitions/TagWithoutJoins"
            }
          },
          "400": {
            "description": "Invalid Tag request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Tag not found"
          }
        }
      },
      "delete": {
        "tags": ["tag"],
        "summary": "Delete the tag",
        "description": "",
        "operationId": "deleteTag",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "ID of tag to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Tag not found"
          }
        }
      }
    },
...
    "/slides": {
      "get": {
        "tags": ["slide"],
        "summary": "Get array of slides",
        "description": "",
        "operationId": "getSlides",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Slide"
              }
            }
          }
        }
      },
      "put": {
        "tags": ["slide"],
        "summary": "Update slides",
        "description": "",
        "operationId": "updateSlides",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin token"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SlideDTO"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful updated",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Slide"
              }
            }
          },
          "400": {
            "description": "Invalid Slide request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/slides/{slideId}": {
      "get": {
        "tags": ["slide"],
        "summary": "Find slide by id",
        "operationId": "findSlideById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "slideId",
            "in": "path",
            "description": "id of slide to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Slide"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Slide not found"
          }
        }
      }
    },
...
    "/advertisements/{advertisementId}": {
      "get": {
        "tags": ["advertisement"],
        "summary": "Find advertisement by id",
        "operationId": "findAdvertisementById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "advertisementId",
            "in": "path",
            "description": "id of advertisement to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Advertisement"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Advertisement not found"
          }
        }
      },
      "put": {
        "tags": ["advertisement"],
        "summary": "Update the advertisement",
        "description": "",
        "operationId": "updateAdvertisement",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin token"
          },
          {
            "name": "advertisementId",
            "in": "path",
            "description": "id of advertisement to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Advertisement"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful updated",
            "schema": {
              "$ref": "#/definitions/Advertisement"
            }
          },
          "400": {
            "description": "Invalid Advertisement request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Advertisement not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ProductWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Tag": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "TagResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
    "TagWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Advertisement": {
      "properties": {
        "id": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "link": {
          "type": "string"
        }
      }
    },
    "Slide": {
      "properties": {
        "id": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "link": {
          "type": "string"
        }
      }
    },
    "SlideDTO": {
      "properties": {
        "image": {
          "type": "string"
        },
        "link": {
          "type": "string"
        }
      }
    }
  }
}